requirements.txt
# aespa-workbook 依赖
# 安装: pip install -r requirements.txt

# OCR & Image Processing
opencv-python>=4.8.0
pytesseract>=0.3.10
Pillow>=10.0.0
numpy>=1.24.0

# Document Generation
python-docx>=0.8.11
jinja2>=3.1.2

# Optional: PDF Conversion (choose one)
# docx2pdf>=0.1.8  # Windows/macOS only
# For Linux, use LibreOffice instead
